Patch added to improve resilience to several types of failure.
The following changes have been made:

- Added "Maximum Execution Time" to the configuration tree.

This option when used will ensure that even after a worker has
been terminated, the execution of the command does not complete
until the duration has been completed.

The purpose of this option is to ensure Supervisor does not
consider a worker to contain fatal flaws if it completes it's
task too quickly.

- Added "Timeout" to the configuration tree.

A suspected cause of worker failures are disconnections to
the gearman server.  Although the worker still waits, physically
unplugging your network cable or wireless does not impact the
worker.  This, and other causes for disconnect leaves the worker
as a zombie that Supervisor does not know needs to be restarted.

This timeout helps prevent a gearman server connection
remaining idle for too long.

- Added Server Exception.

When no server can be connected to successfully with `addServer`,
a `ServerConnectionException` is throwed in the `GearmanExecute`
service.  This will be represented at the command line.

If a minimum execution time is present, the script will still
wait before throwing the exception.  To clean the error messages,
the original warnings are supressed.

Consideration for backwards compatibility may be required here.
However given the work typically does not start when a connection
is not established, and that Supervisor treats immediate
terminations as fatal, this may not have a significant negative
(if negative at all) impact on existing systems.

- Added command line configuration options.

Iterations, Minimum Execution Time, and Timeout have been added
to the following commands:

    gearman:worker:execute
    gearman:job:execute

These options override the base configuration.